+++ opendc-core/src/main/kotlin/com/atlarge/opendc/simulator/kernel/Kernel.kt
@@ -25,55 +25,19 @@
 package com.atlarge.opendc.simulator.kernel
 
 import com.atlarge.opendc.simulator.Bootstrap
-import com.atlarge.opendc.simulator.Entity
-import com.atlarge.opendc.simulator.Instant
 
 /**
- * A message based discrete event simulation kernel.
+ * A message-based discrete event simulator (DES). This interface is a factory for creating [Simulation]s using the
+ * provided [Bootstrap] for the model.
  *
- * The kernel is created by bootstrapping some model `M` (see [Bootstrap]) to simulate and controls the simulation by
- * allowing the user to step over cycles in the simulation and inspect the internal state using [Entity.state].
- *
- * A kernel should provide additionally a [KernelFactory] to create new kernel instances given a certain model
- * [Bootstrap].
- *
- * @param M The shape of the model over which the simulation runs.
  * @author Fabian Mastenbroek (f.s.mastenbroek@student.tudelft.nl)
  */
-interface Kernel<out M> {
+interface Kernel {
     /**
-     * The model in which the simulation runs.
-     */
-    val model: M
-
-    /**
-     * The simulation time.
-     */
-    var time: Instant
-
-    /**
-     * The observable state of an [Entity] in simulation, which is provided by the simulation context.
-     */
-    val <E : Entity<S, *>, S> E.state: S
-
-    /**
-     * Step through one cycle in the simulation. This method will process all events in a single tick, update the
-     * internal clock and then return the control to the user.
-     */
-    fun step()
-
-    /**
-     * Run a simulation over the specified model.
-     * This method will step through multiple cycles in the simulation until no more message exist in the queue.
-     */
-    fun run()
-
-    /**
-     * Run a simulation over the specified model, stepping through cycles until the specified clock tick has
-     * occurred. The control is then handed back to the user.
+     * Create a simulation over the given model facilitated by this simulation kernel.
      *
-     * @param until The point in simulation time at which the simulation should be paused and the control is handed
-     * 				back to the user.
+     * @param bootstrap The apply procedure to apply the simulation with.
+     * @return A [Simulation] instance representing the simulation.
      */
-    fun run(until: Instant)
+    fun <M> create(bootstrap: Bootstrap<M>): Simulation<M>
 }
